supplicant: cleanup notify signals for combined properties in supplicant
Certain properties (for example "scanning") are combined from multiple
other properties. So, we want to notify a changed signal, exactly when
something relevant changes. We also may not want to emit a signal while
we are still in the middle of changing multiple properties together.
Only at certain places we want to check and emit the signal.

Simplify the implementation for that by tracking the property value that
we currently expose, and keeping state about when it changes.